+++ .idea/dictionaries/jtr.xml
@@ -1,0 +1,7 @@
+<component name="ProjectDictionaryState">
+  <dictionary name="jtr">
+    <words>
+      <w>crontab</w>
+    </words>
+  </dictionary>
+</component>
