+++ news/templates/hello.html
@@ -10,20 +10,35 @@
         <button class="btn btn-primary">按我取得新聞</button>
         <script>
             $('container').ready(() => {
+                getData()
+            })
+            function getData() {
+                $('#newsBody').find('tr').remove()
                 $.ajax({
                     type: 'GET',
-                    url: 'http://127.0.0.1:8000/api/news',
+                    url: 'http://127.0.0.1:8000/api/news/',
                     dataType: 'json',
                     timeout: 1000,
-                    success: (result) => {
-                        console.log('result', result);
+                    async: false,
+                    success: (results) => {
+                        let str = '';
+                        $('#newsBody').html('');
+                        results.forEach(result => {
+                            str = `
+                                <tr>
+                                    <th><img src="${result.imgUrl}" width="250"></th>
+                                    <th>${result.title}</th>
+                                    <th>${result.time}</th>
+                                    <th>${result.payload}</th>
+                                    <th class="timestamp">${result.timestamp}</th>
+                                </tr>`
+                            $('#newsBody').append(str)
+                        })
+                        setTimeout(getData, 30000)
                     }
                 })
-            })
+            }
 
-<!--            function getData = {-->
-<!--                -->
-<!--            }-->
         </script>
 
         <div class="wrapper">
@@ -32,28 +47,26 @@
                 <table class="table table-hover">
                     <thead class="thead-dark">
                         <tr>
-                            <th scope="col">ID</th>
                             <th scope="col">圖</th>
                             <th scope="col">標題</th>
-                            <th scope="col">多久之前</th>
+                            <th scope="col">間隔</th>
                             <th scope="col">內容</th>
                             <th scope="col">更新時間</th>
                         </tr>
                     </thead>
-                    <tbody>
-                    {% for new in news %}
-                        <tr>
-                            <th scope="col">{{new.id}}</th>
-                            <th><img src="{{new.imgUrl}}" width="250"></th>
-                            <th>{{new.title}}</th>
-                            <th>{{new.time}}</th>
-                            <th>{{new.payload}}</th>
-                            <th>{{new.timestamp}}</th>
-                        </tr>
-                    {% endfor %}
+                    <tbody id="newsBody">
+<!--                    {% for new in news %}-->
+<!--                        <tr>-->
+<!--                            <th scope="col">{{new.id}}</th>-->
+<!--                            <th><img src="{{new.imgUrl}}" width="250"></th>-->
+<!--                            <th>{{new.title}}</th>-->
+<!--                            <th>{{new.time}}</th>-->
+<!--                            <th>{{new.payload}}</th>-->
+<!--                            <th>{{new.timestamp}}</th>-->
+<!--                        </tr>-->
+<!--                    {% endfor %}-->
                     </tbody>
                 </table>
-
             </div
         </div>
